tests/requirements.txt
pytest
pika
requests
python-dotenv
psutil